ml-engine/requirements.txt
torch
transformers
optimum
onnxconverter-common
onnx
sentencepiece
onnxruntime-directml
onnxruntime-tools
pypdf
fastapi
uvicorn
langchain
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
langchain-huggingface
